MUZ-028 #start-progress #comment some refactorings
Former-commit-id: e53b26ee81240e4973bb5b571b7ca748bf236508

+++ platform/src/main/java/ai/enpasos/muzero/platform/run/ActionExtractor.java
@@ -62,7 +62,7 @@ public class ActionExtractor {
             for (Game game : replayBuffer.getBuffer().getGames()) {
                 List<String> valueList = new ArrayList<>();
                 valueList.add("" + count++);
-                valueList.addAll(game.getGameDTO().getActions().stream().map(a -> a.intValue() + "").collect(Collectors.toList()));
+                valueList.addAll(game.getGameDTO().getActions().stream().map(a -> a + "").collect(Collectors.toList()));
 
                 String[] values = valueList.toArray(new String[0]);
